Rename: 정보목록 페이지 파일명 수정
