+++ petefit/src/pages/information/InformationStyles.jsx
